refactor: 完成代码重复冗余清理
✅ 清理完成:
- 删除旧版 enhanced_unified.rs 文件
- 更新 smart.rs 使用新版 unified.rs
- 清理 mod.rs 废弃导出
- 新增统一任务引擎管理器 EnhancedTaskEngineManager.ts
- 新增统一任务引擎Hook useUnifiedTaskEngine.ts
- 新增类型映射工具 type-mappings.ts
- 修正前端精准获客模块类型引用

🔧 技术改进:
- 统一了 ADB 匹配引擎实现
- 消除了重复的匹配逻辑
- 建立了完整的任务引擎架构
- 提供了类型安全的映射工具
- 保持了向后兼容性

🎯 后续优化:
- 可运行 cargo fix 清理 435 个警告
- 可进行功能测试验证新架构

+++ src-tauri/src/commands/strategy_matching.rs
@@ -192,7 +192,7 @@ pub async fn match_hidden_element_by_text(
     // 构造隐藏元素策略配置
     let config = HiddenElementParentConfig {
         target_text: target_text.clone(),
-        max_traversal_depth: max_traversal_depth.unwrap_or(5) as usize,
+        max_traversal_depth: max_traversal_depth.unwrap_or(5),
         clickable_indicators: vec![
             "Button".to_string(),
             "ImageButton".to_string(),
